requirements.txt
django==4.2.16
djangorestframework==3.14
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.0
python-decouple==3.8
drf-yasg==1.21.8
django-celery-beat==2.7.0
